Move common types to separate files
