+++ src/types/studies.ts
@@ -1,0 +1,6 @@
+export type Study = {
+  id: number;
+  slug: string;
+  name: string;
+  home_page_slug: string;
+};
